changed name of the folder 'interfaces' to 'types'. types reorganized
